refactor: app and index to typescript and removed unused js files
